I did stuff and it's all working perfectly now
