+++ firefly/kernel/kernel.cpp
@@ -30,6 +30,7 @@ void log_core_firefly_contributors() {
 }
 
 [[noreturn]] void kernel_main() {
+	log_core_firefly_contributors();
     panic("Reached the end of the kernel");
     __builtin_unreachable();
 }
