+++ firefly/kernel/memory-manager/virtual/virtual.cpp
@@ -1,12 +1,19 @@
 #include "firefly/memory-manager/virtual/virtual.hpp"
 
+#include "firefly/compiler/clang++.hpp"
 #include "firefly/console/console.hpp"
+#include "firefly/limine.hpp"
 #include "firefly/logger.hpp"
 #include "firefly/memory-manager/primary/primary_phys.hpp"
 #include "firefly/panic.hpp"
 #include "libk++/bits.h"
+#include "firefly/memory-manager/primary/buddy.hpp"
 
 namespace firefly::kernel::mm {
+
+USED struct limine_kernel_address_request kernel_address {
+    .id = LIMINE_KERNEL_ADDRESS_REQUEST, .revision = 0, .response = nullptr
+};
 
 frg::manual_box<kernelPageSpace> kPageSpaceSingleton{};
 
@@ -18,10 +25,13 @@ void kernelPageSpace::init() {
     auto pml4 = static_cast<T *>(Physical::must_allocate());
     kPageSpaceSingleton.initialize(pml4);
 
-    kPageSpaceSingleton.get()->mapRange(PAGE_SIZE, GiB(4), AccessFlags::ReadWrite, AddressLayout::Low);
-    kPageSpaceSingleton.get()->mapRange(0, GiB(2), AccessFlags::ReadWrite, AddressLayout::High);
-    kPageSpaceSingleton.get()->mapRange(0, GiB(2), AccessFlags::Readonly, AddressLayout::Code);
-    kPageSpaceSingleton.get()->mapRange(0, GiB(1), AccessFlags::ReadWrite, AddressLayout::PageData);
+    kPageSpaceSingleton.get()->mapRange(PAGE_SIZE, buddy.get_highest_address(), AccessFlags::ReadWrite, AddressLayout::Low);
+    kPageSpaceSingleton.get()->mapRange(0, GiB(4), AccessFlags::ReadWrite, AddressLayout::High);
+
+    for (size_t i = kernel_address.response->physical_base, j = 0; i < kernel_address.response->physical_base + GiB(1); i += PAGE_SIZE, j += PAGE_SIZE)
+		kPageSpaceSingleton.get()->map(j + kernel_address.response->virtual_base, i, AccessFlags::ReadWrite);
+    
+	kPageSpaceSingleton.get()->mapRange(0, GiB(1), AccessFlags::ReadWrite, AddressLayout::PageData);
     kPageSpaceSingleton.get()->loadAddressSpace();
 
     info_logger << "vmm: Initialized" << logger::endl;
